Added south for DB migrations
South is a third-party app that allows us to make DB migrations easier.  If you change a schema it becomes difficult to automatically update the schema in the database.  Syncdb seems to only allow you to create a new table that doesn't already exist, which is problematic if you are adding/removing a column to a table that already has data.

South allows you to use a few new ./manage.py commands to create "migration" scripts which will allow you to add columns on the fly, without having to destroy existing data.