Phone verification upon signup via SMS
Upon registration of an applicant on the site, we send an SMS message to the mobile phone number entered during sign up.  We want the applicant to confirm the phone number so we don't spam random phone numbers.

Also made another version of USPhoneNumberField which accepts the leading + (plus) symbol which Twilio returns as part of the phone number when someone calls/SMS's us.  The US phone number that Twilio sends is of the format +1XXXYYYZZZZ => (XXX)YYY-ZZZZ