Job creation is now tied to an employer
You need to be logged in as an employer to do job creation.  Actually, it is only checking that you are logged in.  Technically, it may let you see the job creation page if you are logged in as an Applicant, but I believe it will fail if you try to create a job.  Probably could do a little more preventive checks here.