Bad merge when I moved the resume migration script
Moved the migration script from its place to a new space in the timeline because we had conflicting migration scripts.  I believe we had two 0010 or 0011 scripts.  But since I moved just the name over, the script itself was a couple of migrations behind.  Just merged it back to the correct state.